Improve chessbase animation of lame oblique moves
Usually oblique moves are direct leaps, and would be animated by making
the piece jump by specifying a height in cbMoveMidZ. Non-jumping oblique
moves, such as those of the Xiangqi Horse or the Grant-Acedrex Gryphon,
can have the straight-line path to their destination run through occupied
squares, which looks rather ugly. This patch calculates the square closest
to the square of origin the move has to make a detour through in order
to reach the destination purely through diagonal and orthogonal sliding.
This is ambiguous in the case of a 2:1 ration, such as for the Knight jump.
In this case the first leg of the move will be taken orthogonally.
If this is not desired, specifying an (unmeasurably small) negative
jump height with cbMoveMidZ can be used to start diagonally.