Support two new types of chess moves
Move generation in the chessbase infra-structure is extended by defining
two new flags. Both of these would enable moves that would otherwise be
rejected to be pushed on a stack for special moves, which a routine
intercepting the standard move generation can afterwards examine, to see
if they should give rise to pushing more moves on the regular move stack.
These move types are (1) moves to an empty square, and (2) moves to a
square occupied by friend or foe. Type (1) can e.g. be useful for
implementing Checkers-like capture, later testing if the jumped-over square
was occupied by a piece of the right player. Type (2) can be used for
identifying activation of own pieces (as in Knight-relay Chess), the first
leg of a multi-capture (e.g. the Chu-Shogi Lion) or (in combination with a
null step) just to locate pieces that need completely separate move
generation.