Allow user-defined flags on special chess moves
Candidate moves put on the 'specials' move stack will now contain the
move descriptor that gave rise to their generation, as an 'x:' field.
This move descriptor contains all flags that were defined for this move
step when defining the corresponding move graph. This allows the use of
bits to which the move generator attaches no meaning as user-defined flags.
These could then indicate which type of post-processing the candidate moves
require, (e.g. locust capture on the square itself, on a jumped-over square,
or induction of new moves in friendly pieces), in variants that have
different kinds of special moves.