Pass piece count to user evaluation in chessbase
The user-supplied evaluation function was called with 'material'
(containing per-piece-type counts) as an argument, but was kept
ignorant on the total piece counts (also calculated by the chess
base evaluation). This made testing for a bare King (to determine
game end by insufficient mating material) rather cumbersome where
it would otherwise have been trivial ('pieceCount[side] == 1').
So pieceCount is now passed as an extra argument in the call to
the user evaluation.