Make animation speed a variable
The function View.Board.cbAnimate now gets the duration of a move
animation passed as an (optional) argument, rather than hard-coding
it as 600 (ms?). The default value is still 600, for compatibility
with existing code. But this patch makes it possible for variant-specific
overrides to manipulate the animation speed, which is for instance
desirable when animating bent multi-captures.