Add support for entry of two-leg moves
A new module 'multi-leg-view.js' is created, containing an alternative
View.Board.xdInput routine. This routine will pay attention to moves
that have a 'via' field in them, and will interpret the second click
as either specifying the .t or .via. If that leaves multiple matching
moves a third click will be requested to indicate the .t.
 It also contains a wrapper for View.Board.cbAnimate, which checks
whether to move to animate is multi-leg (by the presence of the .via
field), and if so, splits the move in seperate legs, each to be
animated in half the time.