New fairy pieces: sized and deformed
Several new pieces are added to the 'fairy' set, generated from existing
ones by deforming them. For some this is just to give them a size that
is compatible with the King height of the set; these are named 'fr-proper-X',
with X = queen, marshal, cardinal, elephant or crowned Rook. These are
intended as replacements, and use the same 2d sprite as the originals.
Other pieces are meant as new types, and have their own 2d sprite.:
vertically sized rook and bishop (fr-small-rook, fr-small-bishop) for
range-limited versions, enhanced knights (fr-nightrider, fr-ferz-knight,
fr-wazir-knight, fr-zebra), a crowned Bishop (fr-crowned-bishop), a rook
with a tunnel running through it (fr-gate'), and a non-royal king (fr-man).
 The .js geometries are placed in the same directory as the piece from
which they were derived, with which they usually also share the material
texture maps.