Fix automatic locust capture
The input procedure will highlight locust victims rather than to-squares
for the second click, but if that click in not ambiguous (i.e. not also
a to-square, and no locust captures of the same victim with another
to-square), it considers the entry complete. (There would be no ' cancel'
at this stage anyway.) But in that case it has to examine the move for
promotion in the action at the locust square, not at the to-square.