New variant: Scirocco
Another 10x10 variant that uses the mult-leg infrastructure, to implement
rifle capture for the Genie and locust capture for Zig and Zag. It also
uses it for implementing move induction by the Dervish and Harpy.
The baring rule is implemented by keeping track of the number of non-royals
in this.king[2] and [-2], updated in ApplyMove. The Quick(Un)Applies do not
update it; instead the check test relies on a 'tunnel parameter' passed from
QuickApply indicating whether the move was a capture, so it can recognize
situations where the King gets bared, and allows the opponent King to step
into check for that purpose.