Allow custom evaluation of perpetuals in chess base model
When the chess base evaluation detects the maximum number of repetitions
is reached, it now calls a function cb.PerpEval(this,aGame) to determine
the game result (which the function should return), if the game-specific
model has defined one. This way games with complex repetition rules
can take the result determination 'in their own hands'. If the function
is not defined the result is still determined in the old way, based on
the parameter cbOnPerpetual.