Allow custom evaluation of checkmates in chess base model
If the game-specific model has defined a function Model.Game.cbMateEval,
this function will be called when a checkmate is detected, and the
normal game result will be multiplied by what the function return,
to provide the game result. This can be used in variants that require
additional conditions to be fulfilled (which the function can then test).
E.g. in Shogi it can be used to invert the result after a Pawn-drop mate.