Count number of consecutive checks in chess base
The variable this.check now is used as a counter rather than a boolean,
to keep track of the number of consecutive turns the current side to
move has been in check. (This is upward compatible with the old usage.)
A new variable this.oppCheck keeps track of the same for the other
player.
  This patch breaks the kludge to flip a mate result by assigning -1
to this.check, which had to be reverted. (Nothing was using that yet.)