New sub-model: support for drop moves in chess base
By including the modules drop-model.js and drop-view.js (after the
grid-geo-model.js and the selected piece set view, respectively)
some new features are added to the game API:
* Mode.Game.cbDropGeometry(files, ranks, bottomHoldings)
  will provide a proper board of the requested files and ranks,
  with on both sides two extra columns for hand-held pieces,
  and 'bottomHoldings' extra rows at the bottom and top of the
  board for the same purpose. It returns the geometry.
  (The coordinate notation will only appear along the proper board.)
* Mode.Game.cbSetPawnLimit(n) will limit the number of Pawns
  in the same file to n. (Default would be no limit.) It must be
  called from the main body of the game-specific model file.
* Mode.Game.cbDropGraph(geometry, leaperSteps, sliderSteps, bottom, top)
  will return a movement graph for a piece that leaps and slides in
  any of the mentioned directions (passed as arrays of 2-vectors),
  and in addition can be dropped on empty squares of the proper board
  from any holdings square. The parameters 'top' and 'bottom' indicate
  how many ranks are inaccessible by drops at the top and the bottom
  of the board, and by default (i.e. when omitted) are both 0.
* Model.Game.cbAddHoldings(geometry, gameDefinition)
  must be called to set up the holdings, and extend the game definition
  with the pseudo-pieces that will be used to display counters for the
  the number of pieces of a given type in hand, when this gets > 2.
  So instead of returning the game definition directly from cbDefine()
  in the game-specific model file, we must return this.cbAddHoldings()
  acting on that definition.
* View.Game.cbAddCounters(pieces, pieceStyle)
  must be called from the game-specific view file to supplement the
  piece images with the had-counter pseudo-pieces. It accepts what
  otherwise would have been written directly in the 'pieces' property
  cbDefineView() return, plus the name of the 3D piece style we chose
  as piece-set module. What it returns then becomes the 'pieces'
  property. *** BEWARE: *** The 2d images of the counters have to be
  provided as the first 10 clippings of the file with 2d piece images,
  and are assumed to be 100x100 pixels!

The holdings are implemented as a pair of squares in the holdings area
for each piece type: a primary square from which such a piece can be
dropped, and a 'spare' square where a second image of the piece will
be displayed if we have more than one, plus a counter when we have more
than two. Only the counter pseudo-piece is truly on the board; the
pieces in Jocly are displayed from the piece list, so that several of
them can indicate the same square (on which they are then all super-
imposed). But the board can point back to only one of them (the counter),
while the others are put in a linked list dangling from the primary
holdings square. Pawn-per-file counts are hidden in the this.kings array.