Fix repetition detection in chess base model
Repetitions could be overlooked in the chess base model, because
different copies of identical pieces (same type and color) were
considered different, so that swapping their location would not be
recognized as the same position. This because the position hash was
based on the index of the piece in the piece list, rather than on its
(colored) type. In normal chess such swapping is very unlikely, but
in variants where captured pieces can be dropped back it happens quite
often.
  This fix bases the position hash on type, color and location only,
ignoring the piece index. The old API function based on index and
square now accesses the piece of the given index to get its type and
color, and uses 3*type+color instead of the index as first selector
of the key seed, in 'board' mode. This makes the use of 'type' mode
superfluous, so that this is changed in a no-op.
  A more efficient game API call Model.Game.cbBoardKeyChange() is
provided, which directly accepts type, color and square as argument,
and returns the corresponding 'seed' that the caller has to XOR with
the position key. In the base-model code itself the expression returned
by this API is in-lined.
  To make this possible the new implementation keeps the 'seed' tables
in arrays local to base-model.js. Rather than using seed[coltyp][sqr]
it synthesize keys on the fly as typeKey[coltyp]*boardKey[sqr];
to give this a homogeneous distribution of seeds the typeKey is always
odd, while the boardKey is assured to be non-zero by settin some high-
order bit.
  This patch runs the risk of breaking some of the existing chess-
variant implementations that directly interact with the Zobrist API
in a non-standard way. Adding or removing a piece with zobrist.update
calls in 'board' mode should provide no problem, provided the piece's
type and color properties are valid before the call.