Revive old Zobrist hashing scheme
Replacing the flawed hashing scheme broke the opening books, which tabulate
moves by position hash, and cannot be easily converted to a new hashing
scheme. To allow the games that have an opening book (Classic Chess and
Xiangqi) to still use it they will have to use the old hashing scheme.
To this end the 3 different types of key requests (board, piece type and
side to move) are now done by 3 separate routines. A new parameter 'zobrist'
in the structure returned by the cbDefine function in the game's model file
can decide which functions are chosen for this; with a value "old" the
routines emulate the old calls to zobrist(), for which they are just
wrappers. In other cases the new scheme will be used.