Refine selection of special moves
The move modality flag FLAG_CAPTURE_SELF now only pushes the move on
the special stack if the victim is a friend. To push a move when the
victim is an enemy, a new flag FLAG_SPECIAL_CAPTURE in introduced.
By delegating the color test to the base model the custom code in
the game definitions can be simplified.