Pass rider step to special moves in chess module
The move generator in the chess module now encodes the latest step of a
rider move in the moves it puts on the special move stack. For efficiency
reasons this is done by XORing the previously visited (or starting) square
with the current destination element of the graph passed as move.x. This
allows custom code that would be interested in this to retrieve the step
by XORing move.x with move.t.
  The move type can now also define a field antiTrade, which by default
will get value 0 when copies to the pTypes table.
  A third addition is that InitGame now calls an optional routine extraInit,
with geometry as argument. This allows sub-models to do some additional
initialization after the geometry has been initialized.