New sub-model: support common fairy and multi-leg moves in chessbase
This sub-model, when used in combination with base-model.js, offers three
kinds of new support:
1) Some Graph functions for common fairy pieces: Camel, Zebra, Alibaba,
Champion, Wizard, Griffon, Rhino. The latter two are bent sliders, and
a general Graph function cbSkiSlide is added to facilitate adding this
kind of piece, where the first step can have other flags than the remaining
part of the path.
2) Anti-trade rules: piece definitions can now have a field antiTrade,
which specifies the number of an 'anti-trade group'. Pieces are not allowed
to capture a protected piece in the same trade group. Furthermore, when the
group number is negative, there also applies a 'counter-strike' rule,
and you cannot capture a piece in the group if one was captured on the
immediately preceding move.
3) Multi-leg moves: some new mode flags are defined, for hit & run capture,
rifle capture, and checker capture. These all will cause pushing of the
indicated move on the special-moves stack, but the standard move generation
is extended to recognize those there, and finish the processing.
In particular, hit & run moves will indicate a capture that must be followed
by another King step in arbitrary direction. Rifle capture will create a
back-and-forth move to the target, and checker capture (to be used only with
leaps of length 2) capture a jumped-over enemy. The multi-leg moves only
work when the multi-leg-view.js file is included in the view build script.