Support burning in fairy-move-model.js
A new FLAG_BURN is added to the move-modality constants in the FLAG_FAIRY
group. When used with FLAG_HITRUN it causes all enemies adjacent to the
destination to be removed as an automatic side effect of the move. The
extension and effect of this 'burn zone' can be controlled by assigning
a move graph to Model.Game.burnZone, and by setting FLAG_CAPTURE_SELF
as mode in this graph the piece can also be made to burn friendly pieces.
  This patch also makes the hit & run capture pay attention to the mode
flags specified in the Model.Game.neighbors graph, indicating whether
the extra move away from the capture square can be to an empty square
(FLAG_MOVE), capture an enemy (FLAG_CAPTURE) or back to the move's
starting square (FLAG_RIFLE).