Fix Advancer (and improve flag names) in fairy-move-model.js
The Model.Game.cbAdvancerGraph() function forgot to return the graph,
and the Advancer would capture the friendly piece with number 0.
It was a bit confusing that single bits in the move-mode word had the
same name as the fields of Model.Game.cbConstants, which often contain
combinations of bits. Insstead of a FLAG_ prefix the individual bits
now have a _BIT suffix.