Generalize e.p. capture in chessbase
Rather than testing for moves from epCatch pieces to the single epTarget
square, the test generates a line of squares extrapolating backwards from
the prospective e.p. victim and the indicated epTarget, all the way until
the square of origin of the previous move is reached, and compares with
all of those. Note that this would not work for lame oblique moves of
the e.p. victim, but would crash by straying off board.