Support brouhaha squares in chessbase
The chessbase confinement mechanism can now also be used for defining
'brouhaha squares', which are squares that disappear (i.e. become
inaccessible and unpassable) when evacuated. This can be done by
specifying a 'b' for the brouhaha squares in the array that is passed
as the confine parameter of the graph functions. Note that this is
implemented through the functions cbLongRangeGraph and cbShortRangeGraph,
which are called by most graph functions for the standard pieces.
Graph functions that directly create the graph (such as Pawn, Horse or
XQElephant) would ignore this.