Subtle change in how chessbase performs castling
In cbApplyCastle() the King and Rook destinations were taken directly from
the castling table, while the former was also available in the move.t field.
Which was already copied from the castling table during move generation.
  Now it uses the move.t field instead. This provides more opportunity for
custom-generated moves to specify non-standard castlings that differ from
the tabulated castling for the same king and rook location. Especially since
the high-order bits of move.t (which for normally generated castling would
be 0) are now added to the tabulated Rook destination. So it has become
possible to completely control the end-result of the castling through
move.t.