Allow custom perp detection in chessbase to defer judgement
The custom routine cbPerpEval, called in case a position occurs for
the cbMaxRepeats-th time, can now return undefined instead of the game
result. In that case the game will continue. This is useful for
implementing games where the ban on repetition depends on the move
that causes the repeat (e.g. to exempt check evasions).