Support automatic piece definitions in fairy-move-model.js
A routine Model.Game.cbPiecesfFromFEN(geometry,fen,step) is added.
It will return the object with the definitions for the pieces that
occur in the fen, with an 'initial' field that would set those up in
the described position. The 'step' argument indicates the reange of
the initial Pawn move.
  The piece IDs that can be used in the FEN are not only the orthodox
pieces PNBRQK, but also some commonly used fairy pieces: The Wizard (W)
and Champion (H) from Omega Chess, the Archbishop (A) and Marshall (M)
from Capablanca Chess, the Camel (C), (Modern) Elephant (E) and Zebra (Z),
The Xiangqi Cannon (X) and its diagonal counterpart the Vao (V), and
the bent sliders Griffon (G) and Manticore (U).
  Pawns are assumed to be e.p. capable, and Rook are assumed to be
castle pieces.