Improve SkiGraph in fairy-moves-model.js
A larger negative value for the initial flags can now be used to shift
the square where the 2nd leg of the trajectory will start, so that
pieces like D-then-B can be implemented as well. For iflags = -1 the
corner is still skipped (which is good for Ski-sliders and the Grant-
Acedrex Unicorno), but for -2 two steps along the initial vector will
be made before the trajectory bends, etc. The corner then gets the same
flags as the rest of the path.