Make loss of all royals lose in chessbase
The chess base model applies the checking rule, and normally would not
allow a piece with isKing property to be captured. For some of the locust
moves this is difficult to enforce, though. As a backup it now also tests
for absence of any royals (as a free side effect of the kingMoved[]
determination), and flags that as a loss in games where the initial
position did contain royals (as seen by the traching of their location
in king[]).