Improve Pawn support in fairy-move-model.js
Te routine to generate Pawn graphs in fairy-move-model.js is now
cbFlexiPawnGraph(geometry, direction, pawnRank, maxPush).
This generates a position-dependent move, where on all ranks up to and
including the specified pawnRank the Pawn can be pushed to halfway the
board, or by maxPush, whichever is lower. By default pawnRank and maxPush
are equal to the board height, leading to a Pawn that can always be pushed
to halfway the board, no matter where it is and whether it has moved yet.
  By specifying pawnRank as the rank where the (white) Pawns start, Pawns
that have moved will be no longer on it, and thus lose the initial push.
  The routine cbPiecesFromFEN now accepts pawnRank and maxPush as 3rd and
4th argumet too; here the default for the pawnRank is the rank where the
(first encountered) white Pawn starts.