Further development of fairy-move-model
The function cbPiecesFromFEN now returns an object that contains both
the pieceTypes object and a standard promote function. This returned
object is now also equiped with methods addPiece(typeDefinition),
setProperty(name, property, value) and addMoves(name, graph) to tweek
the definitions of the automatically generated types, or add pieces
of types that are not provided automatically.
  There also is a new function cbLionGraph(geometry, confine), and the
Lion and Amazon are added to the piece palette (as L and T).
  The aspects of Griffon, Champion and Wizard are altered to the new
pieces in fairy-set-view that carry their name.