Let cbPiecesFromFEN in fairy-move-model geenrate castlings
The object returned by cbPiecesFromFEN now also contains a castle
property defining some default for castling that in most cases
would be useful. It also has a method setCastling(kstep, partnerName)
that can be used to alter the castlings for moving the King by another
amount, or castling with a piece different from Rook.