New fairy pieces: flying sliders
A flying Queen, Rook and Bishop are added to fairy-set-view. They are
derived from the normal Q, R and B, and their 3d representation used the
same maps. These are modified a bit to provide texture for the wings, in
the location that was used for the bottom. (Which is never visible.)
Their mesh file is therefore located in the same directory as the original
piece.