Provide method to alter cbPiecesFromFEN's piece values
The object returned by cbPiecesFromFEN now contains a method setValues(),
which accepts an object with properties names as the piece-type
abbreviations, that as value have the newly desired piece value for that
type. Like p.setValues({P: 0.8, N:2.5, B:4, R:6}) to specify new values for
Pawn, Knight, Bishop and Rook.