Fix animation on non-grid boards
The automatic splitting of slides into an orthogonal and diagonal leg
broke things on hex boards, where the principal directions of motion
are completely different. The involved code is now only executed if
cbMoveMidZ request it. It must furthermore choose itself which of the
legs to animate first: when a z-coordinate is specified 1 unit below the
average height, the orthogonal leg goes first, for 2 units the diagonal leg,
irrespective of their length.